DE/AI/requirements.txt
flask
PyPDF2
google-generativeai
langchain
langchain-community
langchain-google-genai
faiss-cpu
python-dotenv
werkzeug